pyproject.toml
[tool.poetry]
name = "film_watch_site"
version = "0.1.0"
description = ""
authors = ["Abdurashid Khikmatov"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
celery = "^5.2.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
flower = "^1.2.0"
Pillow = "^9.5.0"
psycopg2-binary = "^2.9.6"
django-autoslug = "^1.9.9"
django-redis = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
